requirements.txt
bit
pycryptodome
base58
